pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gpu-llm-infra-lab"
version = "0.1.0"
description = "Tiny GPT training + GPU/collective benchmarks for infra/LLM internship portfolio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "numpy>=1.24.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
onnx = ["onnx>=1.14.0"]

[project.scripts]
gli-train = "gpu_llm_infra_lab.train:main"

[tool.setuptools.packages.find]
where = ["src"]
